// src/features/Header/types/announcementType.d.ts
export type AnnouncementType = {
  community: null;
  communityID: number;
  content: string;
  createdAt: string;
  createdBy: string;
  createdByUser: null;
  expiresAt: string;
  isDeleted: boolean;
  modifiedAt: null;
  modifiedBy: null;
  modifiedByUser: null;
  noticeID: number;
  title: string;
};